requirements.txt
ffmpeg-python==0.2.0
flask-restx==1.3.0
huey==2.5.2
paho-mqtt==2.1.0
phonemizer==3.3.0
styletts2 @ git+https://github.com/sidharthrajaram/StyleTTS2@cf48b82718c38ad872cd3c7189714fe93928ca5f
torchaudio==2.5.1
